requirement.txt
asgiref==3.2.3
bcrypt==3.1.7
boto3==1.14.16
botocore==1.17.16
bson==0.5.8
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dataclasses==0.6
Django==2.2.11
djangorestframework==3.11.0
djongo==1.3.2
docutils==0.15.2
fabric==2.5.0
idna==2.9
invoke==1.4.1
jmespath==0.10.0
paramiko==2.7.1
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.9
